# lib/api-zod/src/generated/types/clinicDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediBridge Global API
 * OpenAPI spec version: 0.1.0
 */
import type { Clinic } from './clinic';
import type { Doctor } from './doctor';

export type ClinicDetail = Clinic & {
  doctors?: Doctor[];
};