# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediBridge Global API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  name: string;
  title: string;
  specialty: string;
  licenseNumber: string;
  yearsExperience: number;
  certifications: string[];
  bio?: string;
  imageUrl?: string;
  languages: string[];
  verified: boolean;
  documentHash?: string;
  onChainTxHash?: string;
  onChainTimestamp?: number;
}