# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediBridge Global API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  name: string;
  location: string;
  procedure: string;
  clinic: string;
  rating: number;
  beforeText: string;
  afterText: string;
  savings: number;
  avatarUrl: string;
  anonymous?: boolean;
  patientWallet?: string;
  documentHash?: string;
  onChainTxHash?: string;
  onChainTimestamp?: number;
  verified?: boolean;
}